requirements.txt
streamlit
pandas
numpy
scikit-learn
textblob
python-binance
matplotlib
openai
python-dotenv
joblib
